requirements.txt
Django==5.1.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-redis==5.4.0
sqlparse==0.5.1
tzdata==2024.1

psycopg2-binary==2.9.9

celery==5.3.6
billiard==4.2.0
kombu==5.3.5
amqp==5.1.1
vine==5.0.1

dlib==19.22.99
face_recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.24.4
pillow==10.2.0
python-dateutil==2.9.0.post0
PyJWT==2.8.0
packaging==24.1
redis==5.0.3
click==8.1.7
asgiref==3.8.1
